Web/requirement.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.28.1
pandas==1.5.3
numpy==1.21.6
fpdf==1.7.2